+++ api/src/main/java/cn/com/cmdd/dao/OrdersDao.java
@@ -15,33 +15,33 @@ public interface OrdersDao {
 	//:+
 	void saveOrders(Orders orders);
 	//-
-	void deleteOrders(@Param("id") Integer id);
-	void deleteCallService(@Param("orderId")Integer orderId);	
-	void deleteOrderItemByDiningTableId(@Param("diningTableId") Integer diningTableId);
-	void deleteCallServiceByDiningTableId(@Param("diningTableId")Integer diningTableId);
-    void deleteMemberLotterLogByOrderId(@Param("orderId")Integer orderId);
+	void deleteOrders(@Param("id") Long id);
+	void deleteCallService(@Param("orderId")Long orderId);	
+	void deleteOrderItemByDiningTableId(@Param("diningTableId") Long diningTableId);
+	void deleteCallServiceByDiningTableId(@Param("diningTableId")Long diningTableId);
+    void deleteMemberLotterLogByOrderId(@Param("orderId")Long orderId);
     //:\
-    void updateIsUpload(@Param("id")Integer id);
+    void updateIsUpload(@Param("id")Long id);
     
 	void updateOrders(Orders Order);
 	void resetOrderPno(@Param("shopId") Integer shopId);
 	void unifyOrderNo(
-			@Param("diningTablePid") Integer pId,
-			@Param("diningTableIdList") List<Integer> idList);
-	void resetDiningTable(@Param("id")Integer id);
+			@Param("diningTablePid") Long pId,
+			@Param("diningTableIdList") List<Long> idList);
+	void resetDiningTable(@Param("id")Long id);
 	//:/
 	List<Orders> selectByIsUpload(); 
 	Orders selectByDiningTableAndStatusId(
-			@Param("diningTableId")Integer diningTableId,
+			@Param("diningTableId")Long diningTableId,
 			@Param("statusId")Integer statusId);
-	List<Integer> selectIdListByPId(@Param("PId") Integer PId);
-	List<Orders> selectListById(@Param("id") Integer id);
+	List<Long> selectIdListByPId(@Param("PId") Long PId);
+	List<Orders> selectListById(@Param("id") Long id);
 	List<Orders> selectList(
             @Param("shop_id") Integer shop_id,
             @Param("id") Integer id,
             @Param("page_size") Integer page_size,
             @Param("pageStart_index") Integer pageStart_index,
-            @Param("dining_table_id") Integer dining_table_id,
+            @Param("dining_table_id") Long dining_table_id,
             @Param("start_time") Date start_time,
             @Param("end_time") Date end_time,
             @Param("statusList") List<Integer> statusList,
